Исправлены баги отображения, добавлен контролькол-во выводимых файлов из конфига

+++ FileManager_Budgo/Program.cs
@@ -20,20 +20,20 @@ namespace FileManager_Budgo
             Console.WriteLine(@"exit-Выход");
         }
 
-        static void CatalogDirs(DirectoryInfo[] subDirs)
+        static void CatalogDirs(DirectoryInfo[] subDirs, int q)
         {
-            if (subDirs.Length >= 10)
+            if (subDirs.Length >= q)
             {
                 int subDirscoll = 0;
                 int i1 = subDirscoll;
                 bool flag = true;
                 do
                 {
-                    if (subDirscoll + 10 >= subDirs.Length)
+                    if (subDirscoll + q >= subDirs.Length)
                     {
                         for (int i = subDirscoll; i < subDirs.Length; i++)
                         {
-                            Console.Write("├─");
+                            Console.Write("├─ ");
                             Console.WriteLine(subDirs[i].Name);
                             subDirscoll++;
                         }
@@ -42,9 +42,9 @@ namespace FileManager_Budgo
                         
                     }
                     else i1 = subDirscoll;
-                    for (int i = subDirscoll; i < i1 + 10; i++)
+                    for (int i = subDirscoll; i < i1 + q; i++)
                     {
-                        Console.Write("├─");
+                        Console.Write("├─ ");
                         Console.WriteLine(subDirs[i].Name);
                         subDirscoll++;
                     }
@@ -58,11 +58,13 @@ namespace FileManager_Budgo
                         {
                             case "да":
                                 flag = true;
+                                Console.Clear();
                                 break;
                             case "нет":
                                 flag = false;
                                 break;
                             default:
+                                Console.Clear();
                                 Console.WriteLine("Введена не правильная команда. Вывод продолжится автомотически.");
                                 flag = true;
                                 break;
@@ -75,26 +77,26 @@ namespace FileManager_Budgo
             {
                 for (int i = 0; i < subDirs.Length; i++)
                 {
-                    Console.Write("├─");
+                    Console.Write("├─ ");
                     Console.WriteLine(subDirs[i].Name);
                 }
             }
         }
 
-        static void Catalogfile(FileInfo[] subfile)
+        static void Catalogfile(FileInfo[] subfile, int q)
         {
-            if (subfile.Length >= 10)
+            if (subfile.Length >= q)
             {
                 int subfilecoll = 0;
                 int i1 = subfilecoll;
                 bool flag = true;
                 do
                 {
-                    if (subfilecoll + 10 >= subfile.Length)
+                    if (subfilecoll + q >= subfile.Length)
                     {
                         for (int i = subfilecoll; i < subfile.Length; i++)
                         {
-                            Console.Write("─");
+                            Console.Write("─ ");
                             Console.WriteLine(subfile[i].Name);
                             subfilecoll++;
                         }
@@ -102,14 +104,14 @@ namespace FileManager_Budgo
                         flag = false;
                     }
                     else i1 = subfilecoll;
-                    for (int i = subfilecoll; i < i1 + 10; i++)
+                    for (int i = subfilecoll; i < i1 + q; i++)
                     {
-                        Console.Write("─");
+                        Console.Write("─ ");
                         Console.WriteLine(subfile[i].Name);
                         subfilecoll++;
                     }
 
-                    if (subfilecoll < subfile.Length - 10)
+                    if (subfilecoll < subfile.Length)
                     {
                         Console.WriteLine("...");
                         Console.WriteLine("Продолжить вывод? дa / нет");
@@ -118,11 +120,13 @@ namespace FileManager_Budgo
                         {
                             case "да":
                                 flag = true;
+                                Console.Clear();
                                 break;
                             case "нет":
                                 flag = false;
                                 break;
                             default:
+                                Console.Clear();
                                 Console.WriteLine("Введена не правильная команда. Вывод продолжится автомотически.");
                                 flag = true;
                                 break;
@@ -135,21 +139,21 @@ namespace FileManager_Budgo
             {
                 for (int i = 0; i < subfile.Length; i++)
                 {
-                    Console.Write("─");
+                    Console.Write("─ ");
                     Console.WriteLine(subfile[i].Name);
                 }
             }
         }
 
-        static void Catalog(DirectoryInfo dir)
+        static void sl(DirectoryInfo dir, int q)
         {
             try
-            {
+            {                
                 Console.WriteLine(dir.Name);
                 DirectoryInfo[] subDirs = dir.GetDirectories();
                 FileInfo[] subfile = dir.GetFiles();
-                CatalogDirs(subDirs);
-                Catalogfile(subfile);
+                CatalogDirs(subDirs, q);
+                Catalogfile(subfile, q);
             }
             catch (DirectoryNotFoundException dirEx)
             {
@@ -165,11 +169,11 @@ namespace FileManager_Budgo
         /// </summary>
         /// <param name="dirName">Текущая директория</param>
         /// <param name="Cat">Каталог</param>
-        static string cdCatalog(string dirName, string cat)
+        static string cd(string dirName, string cat, int q)
         {
             string nevDir = dirName + @"\" + cat;
             DirectoryInfo newDir = new DirectoryInfo(nevDir);
-            Catalog(newDir);
+            sl(newDir, q);
             return nevDir;
         }
 
@@ -180,6 +184,7 @@ namespace FileManager_Budgo
                 Console.WriteLine("Введите новое название файла: ");
                 string newcat = Console.ReadLine();
                 CopyDirectory(Path.Combine(dirName, cat), Path.Combine(dirName, newcat));
+                Console.WriteLine("Копирование произведено успешно");
             }
             catch (DirectoryNotFoundException)
             {
@@ -214,12 +219,20 @@ namespace FileManager_Budgo
         {
             try
             {
-                if (File.Exists(Path.Combine(Dir, file)))
+                if (file == "")
                 {
-                    File.Delete(Path.Combine(Dir, file));
-
+                    Console.WriteLine("Файл не был указан");
                 }
-                else Directory.Delete(Path.Combine(Dir, file), true);
+                else
+                {
+                    if (File.Exists(Path.Combine(Dir, file)))
+                    {
+                        File.Delete(Path.Combine(Dir, file));
+
+                    }
+                    else Directory.Delete(Path.Combine(Dir, file), true);
+                    Console.WriteLine("Удаление произведено успешно!");
+                }
             }
             catch (DirectoryNotFoundException)
             {
@@ -234,11 +247,13 @@ namespace FileManager_Budgo
         static void crfile(string Dir, string file)
         {
             File.Create(Path.Combine(Dir, file));
+            Console.WriteLine("Файл создан");
         }
 
         static void crcat(string Dir, string file)
         {
             Directory.CreateDirectory(Path.Combine(Dir, file));
+            Console.WriteLine("Каталог создан");
         }
 
         static void info(string Dir, string file)
@@ -267,17 +282,37 @@ namespace FileManager_Budgo
             }
         }
 
+        static void configSave(string config, string folderPath)
+        {
+            string[] Config = File.ReadAllLines(config);
+            Config[0] = folderPath;
+            File.WriteAllLines(config, Config);
+        }
+        static string configDirecvori(string config)
+        {
+            string[] configDirmas = File.ReadAllLines(config);
+            return configDirmas[0];
+        }
+        static int configpar(string config)
+        {
+            string[] configDirmas = File.ReadAllLines(config);
+            return Convert.ToInt32(configDirmas[1]);
+        }
+
         static void Main(string[] args)
         {
             string memu;
             bool end = false;
             string config = "Config.txt";
             string configDir = @"C:\";
+            int q = 0;
             help();
             do
-            {                
+            {
+                Console.WriteLine("Введите команду");
                 memu = Console.ReadLine();
-                switch (@memu)
+                Console.Clear();
+                switch (memu)
                 {
                     case "help":
                         help();
@@ -286,50 +321,52 @@ namespace FileManager_Budgo
                         Console.WriteLine("Введите путь к каталогу: ");
                         string folderPath = Console.ReadLine();
                         DirectoryInfo dir = new DirectoryInfo(folderPath);
-                        File.WriteAllText(config, folderPath);
-                        Catalog(dir);
+                        configSave(config, folderPath);
+                        q = configpar(config);
+                        sl(dir,q);
                         break;
                     case "cd":
                         Console.WriteLine("Введите название каталога: ");                        
-                        string newCat = Console.ReadLine();
-                        configDir = File.ReadAllText(config);
-                        File.WriteAllText(config, cdCatalog(configDir, newCat));
+                        string newCat = Console.ReadLine();                      
+                        configDir = configDirecvori(config);
+                        q = configpar(config);
+                        configSave(config, cd(configDir, newCat, q));
                         break;
                     case "rm":
                         Console.WriteLine("Введите название файла или каталога для удаления: ");
                         string rmCat = Console.ReadLine();
-                        configDir = File.ReadAllText(config);
+                        configDir = configDirecvori(config);
                         rm(configDir, rmCat);
                         break;
                     case "cp":
                         Console.WriteLine("Введите название файла или каталога для копирования: ");
                         string cpCat = Console.ReadLine();
-                        configDir = File.ReadAllText(config);
+                        configDir = configDirecvori(config);
                         cp(configDir, cpCat);
                         break;
                     case "crfile":
                         Console.WriteLine("Введите название файла для создания ");
                         string crFile = Console.ReadLine();
-                        configDir = File.ReadAllText(config);
+                        configDir = configDirecvori(config);
                         crfile(configDir, crFile);
                         break;
                     case "crcat":
                         Console.WriteLine("Введите название каталога для создания ");
                         string crCat = Console.ReadLine();
-                        configDir = File.ReadAllText(config);
+                        configDir = configDirecvori(config);
                         crcat(configDir, crCat);
                         break;
                     case "file":
-                        configDir = File.ReadAllText(config);
+                        configDir = configDirecvori(config);
+                        q = configpar(config);
                         DirectoryInfo dirfile = new DirectoryInfo(configDir);
-                        Catalog(dirfile);
+                        sl(dirfile, q);
                         break;
                     case "info":
-                        configDir = File.ReadAllText(config);
+                        Console.WriteLine("Введите название файла или каталога: ");
                         string infoCat = Console.ReadLine();
-                        configDir = File.ReadAllText(config);
+                        configDir = configDirecvori(config);
                         info(configDir, infoCat);
-
                         break;
                     case "exit":
                         end = true;
